seperated command loading vs handling
